# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UniHER API",
    "description": "API da plataforma UniHER - Saúde e bem-estar corporativo feminino",
    "version": "1.0.0",
    "contact": {
      "name": "UniHER",
      "url": "https://uniher.com.br"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor principal"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Cookie httpOnly com token JWT"
      }
    },
...
  },
  "paths": {
...
    "/auth/reset-password": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Redefinir senha",
        "description": "Define nova senha usando token de redefinição",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "password"],
                "properties": {
                  "token": { "type": "string" },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Senha redefinida com sucesso" },
          "400": { "description": "Token inválido ou expirado" }
        }
      }
    },
...
    "/admin/users/{id}": {
      "get": {
        "tags": ["Admin - Usuários"],
        "summary": "Detalhes do usuário",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Dados do usuário" }
        }
      },
      "put": {
        "tags": ["Admin - Usuários"],
        "summary": "Atualizar usuário",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Usuário atualizado" }
        }
      },
      "delete": {
        "tags": ["Admin - Usuários"],
        "summary": "Remover usuário",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Usuário removido" }
        }
      }
    },
...
    "/gamification/daily-missions/{id}/complete": {
      "post": {
        "tags": ["Gamificação"],
        "summary": "Completar missão diária",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Missão completada" }
        }
      }
    },
...
    "/company": {
      "get": {
        "tags": ["Empresa"],
        "summary": "Dados da empresa do usuário",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Dados da empresa" }
        }
      }
    },
...
    "/rh/objectives/{id}": {
      "put": {
        "tags": ["RH - Objetivos"],
        "summary": "Atualizar objetivo",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Objetivo atualizado" }
        }
      },
      "delete": {
        "tags": ["RH - Objetivos"],
        "summary": "Remover objetivo",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Objetivo removido" }
        }
      }
    },
...
    "/push/subscribe": {
      "post": {
        "tags": ["Push Notifications"],
        "summary": "Registrar subscrição push",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Subscrição registrada" }
        }
      }
    }
  },
...
}